Warn about broken links on WP entry header
People keep complaining in #freenode about links on decade-old blog posts being broken. One "easy win" for this is warning about that in the header; long-term there should be a redirect or old posts should be edited to point to an informative interstitial page, but this is better than nothing.